[Refactor][전해리] DataProcessing 파일 정리
